powerstat: handle cases where -ve deltas occur from /proc/stat
Sometimes we get -ve deltas on the data from /proc/stat, this
has been observed on Nexus 4 phones.  So guard against this by
forcing a zero return rather than a -ve result.  Potentially all
the values could go -ve, which means we need to also guard that the
total time is not zero.  In these error conditions return false
and re-sample.

Signed-off-by: Colin Ian King <colin.king@canonical.com>