powerstat: fix leak of info when there are no free slots in table
Coverity discovered a memory leak of info when there are no
free slots in the proc_info table.  Instead of allocating info
and then trying to add it to a full proc_info table we should instead
try and find a free slot and bail out early if none are available.

Signed-off-by: Colin Ian King <colin.king@canonical.com>