Add thread/child detection, add better proc name fetching
Now detect if a new process is a child or a thread and print this
state out.  Remove the tpid info, it is extraneous.

Improve the cmdline process info gathering, fallback to getting
info from comm field (for example, kernel theads don't have a cmdline)
and if we can't get the comm field then show that the data is unknown
rather than have an empty field.

Finally, run at highest priority so we can potentially catch short
lived processes. This is not perfect, since proc information is racy
but it is better than nothing.

Signed-off-by: Colin Ian King <colin.king@canonical.com>